requirements.txt
pandas
tabula-py>=2.3.0
PyPDF2>=2.0.0
numpy
openpyxl
